Generic commands for smart instead of arch specific
The SMART and other commands will need special treatment with time and
only require the arch specific code to pass a CDB for SCSI commands (or
ATA PASSTHROUGH SCSI commands for SMART).

This is as opposed to the possibly posix variants to read/write/inquiry
and read capacity that may not be a SCSI device, especially if we end
up testing an SD card or some other strange disk type (possibly also old
PATA).